Review Parser - Remove phone number from formatted address field.
By removing the phone number from the formatted address field we can drastically increase our chances of success when querying Google for location based information. This also has the added benefit of providing us with the ability to extract the phone number for a restaurant and adding it to our model if we so wish.